Refactor service control and fix scheduler
