+++ app/src/main/java/com/d4rk/lowbrightness/receivers/OnBootBroadcastReceiver.kt
@@ -3,7 +3,7 @@ package com.d4rk.lowbrightness.receivers
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.d4rk.lowbrightness.base.Application.refreshServices
+import com.d4rk.lowbrightness.base.ServiceController.refreshServices
 
 class OnBootBroadcastReceiver : BroadcastReceiver() {
     override fun onReceive(context : Context , intent : Intent) {
